Cleanup pull request template
This updates four sections:

- In the "Screenshots" section it removes the "if appropriate" aside.
Every section in this template should be included only if it's
appropriate. There is no need to call out this section in particular.

- The "Checklist" section was removed. After nearly two years of using
this template, I still haven't seen anybody use this section
correctly. The intention was that the reviewer would use this section to
ensure that each pull request met a baseline standard. Instead, the
author tends to check a few boxes at random. This provides.

- The "Release Type" section was removed. This is useful when authoring
libraries but the vast majority of the work we do is application- or
script-based. Almost 100% of the time, a random value of either "patch"
or "minor" is selected, regardless of what the actual semantic version
release is. Given that, this section provides little value.

- The "Manually Tested" section now comes pre-populated with a GitHub
Flavored Markdown select option. We use these to allow the reviewer to
check off the test steps as they review (much like the original
intention with the "Checklist" section). GitHub will display the number
of checked select options as part of the pull request preview. This view
was always useless to us because of how many unchecked select options we
had between the "Checklist" and "Release Type" sections. Now that those
are gone, the pull request preview view will provide more useful info.